chore: get clippy and fmt clean on workspace (CI was failing)
The new ci.yml (added in b788102) runs fmt + clippy -D warnings on
every PR — first runs surfaced ~25 unique pre-existing lint and
format issues that had accumulated without a gate.

Auto-applied via tooling:
- cargo fmt --all (mostly main.rs + prefs.rs, with cascading touches
  in 23 other files from clippy's secondary fmt pass)
- cargo clippy --fix for derivable_impls, needless_lifetimes,
  redundant_closure, needless_borrow, match_single_binding, map_entry,
  manual_range_contains, manual_inspect, collapsible_if,
  blocks_in_conditions, unnecessary_cast, unnecessary_map_or, and
  some doc_lazy_continuation + ptr_arg cases.

Manual fixes:
- edge.rs:119 — let-else → ? operator (clippy::question_mark)
- main.rs:802 — unwrap-after-is_some replaced with .and_then() chain
  (clippy::unnecessary_unwrap)
- hud_render.rs:2027 — Stroke::default() + reassign collapsed into
  struct literal (clippy::field_reassign_with_default)
- types.rs / main.rs — reflowed doc comments so leading "+" in a wrapped
  line doesn't get parsed as a markdown bullet, and added a blank
  /// line where a list runs into a paragraph (clippy::doc_lazy_continuation)
- prefs.rs:1378 — removed trailing whitespace rustfmt couldn't resolve
- main.rs apply_nudge_step + handle_pointer_button — Vec params
  converted to slices where the function body didn't need .push/.remove
  (clippy::ptr_arg)

#[allow] for structural lints we won't address as part of this cleanup:
- clippy::too_many_arguments on 8 functions (8-18 args). Real refactor
  candidates but out of scope for a CI-fix commit; rendering / event
  hot paths intentionally take many args to avoid intermediate structs.
- clippy::large_enum_variant on wayland.rs:315 enum Cmd. Boxing the
  largest variant ripples through every match arm; skip for now.
- clippy::ptr_arg on handle_pointer_button held_rects, which the body
  does .push/.remove on.

Verified: cargo fmt --check, cargo clippy --workspace --all-targets
--all-features -- -D warnings, and cargo test all pass locally.

+++ crates/vernier-platform/src/linux/screencast.rs
@@ -145,7 +145,10 @@ fn run_pipewire(
     }
 
     let _ = ready_tx.send(Ok(()));
-    log::info!("pipewire: main loop running with {} stream(s)", streams_meta.len());
+    log::info!(
+        "pipewire: main loop running with {} stream(s)",
+        streams_meta.len()
+    );
 
     mainloop.run();
     Ok(())
@@ -266,7 +269,7 @@ fn create_stream(
                         mmap_len,
                         libc::PROT_READ,
                         libc::MAP_PRIVATE,
-                        dfd as i32,
+                        dfd,
                         0,
                     )
                 };
@@ -349,7 +352,7 @@ fn build_buffers_pod() -> Result<Vec<u8>> {
     let prop = Property {
         key: pw::spa::sys::SPA_PARAM_BUFFERS_dataType,
         flags: PropertyFlags::empty(),
-        value: Value::Int(mem_types as i32),
+        value: Value::Int(mem_types),
     };
     let obj = Object {
         type_: SpaTypes::ObjectParamBuffers.as_raw(),
@@ -422,18 +425,19 @@ pub(crate) fn open_session_blocking() -> Result<SessionState> {
         if let Err(e) = save_token(token) {
             log::warn!("screencast: could not persist restore token: {e}");
         } else {
-            log::info!("screencast: persisted restore token at {}", token_path_display());
+            log::info!(
+                "screencast: persisted restore token at {}",
+                token_path_display()
+            );
         }
     }
     Ok(result)
 }
 
 async fn open_session_async(prev_token: Option<String>) -> Result<SessionState> {
-    let proxy = Screencast::new()
-        .await
-        .map_err(|e| PlatformError::Portal {
-            reason: format!("create screencast proxy: {e}"),
-        })?;
+    let proxy = Screencast::new().await.map_err(|e| PlatformError::Portal {
+        reason: format!("create screencast proxy: {e}"),
+    })?;
 
     let session = proxy
         .create_session(Default::default())
